backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-bcrypt
flask-mail
flask-limiter
pandas
numpy
scikit-learn
pymongo
requests
python-dotenv
redis
apscheduler
psutil
gunicorn
matplotlib==3.8.3
seaborn==0.13.2
fpdf